♻::#15 도메인 엔티티 리펙토링
