+++ src/main/java/com/newfangled/flockbackend/domain/project/entity/BoardComment.java
@@ -1,22 +1,28 @@
-package com.newfangled.flockbackend.domain.team.entity;
+package com.newfangled.flockbackend.domain.project.entity;
 
+import com.newfangled.flockbackend.domain.account.entity.Account;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import javax.validation.constraints.NotNull;
 
 @Getter
 @AllArgsConstructor @NoArgsConstructor
 @Builder
 @Entity
-public class Team {
+public class BoardComment {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @NotNull
-    private String name;
+    @OneToOne
+    private Board board;
+
+    @OneToOne
+    private Account account;
+
+    private String content;
+
 }
