+++ src/main/java/com/newfangled/flockbackend/domain/project/entity/Project.java
@@ -1,28 +1,26 @@
 package com.newfangled.flockbackend.domain.project.entity;
 
-import com.newfangled.flockbackend.domain.project.embed.TeamId;
-import com.newfangled.flockbackend.domain.team.entity.TeamMember;
+import com.newfangled.flockbackend.domain.team.entity.Team;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.EmbeddedId;
-import javax.persistence.Entity;
-import javax.persistence.ManyToOne;
+import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 
 @Getter
 @AllArgsConstructor @NoArgsConstructor
+@Builder
 @Entity
 public class Project {
-    
-    @EmbeddedId
-    private TeamId teamId;
 
-    @ManyToOne
-    private TeamMember teamMember;
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @OneToOne
+    private Team team;
 
     @NotNull
     private String name;
-    
 }
