+++ src/main/java/com/newfangled/flockbackend/domain/sns/repository/CommentRepository.java
@@ -1,10 +1,10 @@
 package com.newfangled.flockbackend.domain.sns.repository;
 
-import com.newfangled.flockbackend.domain.sns.entity.Comment;
+import com.newfangled.flockbackend.domain.sns.entity.SNSComment;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentRepository extends CrudRepository<Comment, Long> {
+public interface CommentRepository extends CrudRepository<SNSComment, Long> {
 
 }
