+++ src/main/java/com/newfangled/flockbackend/domain/team/entity/TeamMember.java
@@ -1,6 +1,8 @@
 package com.newfangled.flockbackend.domain.team.entity;
 
-import com.newfangled.flockbackend.global.embed.TeamPosition;
+import com.newfangled.flockbackend.domain.account.entity.Account;
+import com.newfangled.flockbackend.domain.team.type.Role;
+import com.newfangled.flockbackend.global.embed.TeamId;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -15,10 +17,12 @@ import javax.persistence.*;
 @Table(name = "team_member")
 public class TeamMember {
 
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    @EmbeddedId
+    private TeamId teamId;
 
-    @Embedded
-    private TeamPosition teamPosition;
+    @OneToOne
+    private Account account;
 
+    @Enumerated(EnumType.STRING)
+    private Role role;
 }
