+++ src/main/java/com/newfangled/flockbackend/domain/team/type/Role.java
@@ -1,5 +1,5 @@
 package com.newfangled.flockbackend.domain.team.type;
 
 public enum Role {
-    LEADER, SUB_LEADER, PM, MEMBER      // 팀장, 부팀장, PM, 팀원
+    Member, Leader, ForeignMember
 }
