+++ src/main/java/com/newfangled/flockbackend/global/embed/BoardId.java
@@ -1,0 +1,17 @@
+package com.newfangled.flockbackend.global.embed;
+
+import com.newfangled.flockbackend.domain.project.entity.Board;
+import lombok.Getter;
+
+import javax.persistence.Embeddable;
+import javax.persistence.ManyToOne;
+import java.io.Serializable;
+
+@Getter
+@Embeddable
+public class BoardId implements Serializable {
+
+    @ManyToOne
+    private Board board;
+
+}
