+++ src/main/java/com/newfangled/flockbackend/global/jwt/dto/request/TokenRequestDto.java
@@ -1,0 +1,4 @@
+package com.newfangled.flockbackend.global.jwt.dto.request;
+
+public class TokenRequestDto {
+}
